Added a check to prevent the user from deleting themself

+++ WebApi/Features/Users/Commands/DeleteUser.cs
@@ -1,5 +1,6 @@
 using Helpers.Framework.Exceptions;
 using MediatR;
+using WebApi.Core.Auth.Services;
 using WebApi.Infrastructure.Persistence;
 using WebApi.Infrastructure.Persistence.Entities;
 
@@ -15,18 +16,23 @@ public static class DeleteUser
     internal sealed class Handler : IRequestHandler<Command>
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserAccessor _userAccessor;
 
-        public Handler(ApplicationDbContext context)
+        public Handler(ApplicationDbContext context, UserAccessor userAccessor)
         {
             _context = context;
+            _userAccessor = userAccessor;
         }
 
         public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
         {
             var user = await _context.Users.FindAsync(request.Id, cancellationToken);
+
             if (user == null)
                 throw new NotFoundException(nameof(User), request.Id);
 
+            if (user.Id == _userAccessor.User.Id)
+                throw new BadRequestException("You cannot delete yourself");
 
             _context.Users.Remove(user);
             await _context.SaveChangesAsync(cancellationToken);
